+++ src/main/java/com/kustaurant/kustaurant/api/restaurant/TierApiController.java
@@ -2,6 +2,8 @@ package com.kustaurant.kustaurant.api.restaurant;
 
 import com.kustaurant.kustaurant.common.restaurant.domain.enums.LocationEnum;
 import com.kustaurant.kustaurant.common.restaurant.infrastructure.restaurant.RestaurantEntity;
+import com.kustaurant.kustaurant.common.restaurant.service.RestaurantApiService;
+import com.kustaurant.kustaurant.common.restaurant.service.TierService;
 import com.kustaurant.kustaurant.global.UserService;
 import com.kustaurant.kustaurant.global.apiUser.customAnno.JwtToken;
 import com.kustaurant.kustaurant.global.exception.exception.ParamException;
@@ -44,6 +46,8 @@ import java.util.stream.IntStream;
 @RequiredArgsConstructor
 public class TierApiController {
 
+    private final TierService tierService;
+
     private final RestaurantApiService restaurantApiService;
     private final UserService userService;
     private final RestaurantFavoriteService restaurantFavoriteService;
@@ -83,7 +87,7 @@ public class TierApiController {
         // page 0부터 시작하게 수정
         page--;
         // DB 조회
-        List<RestaurantEntity> restaurants = restaurantApiService.getRestaurantsByCuisinesAndSituationsAndLocationsWithPage(cuisines, situations, locations, null, true, page, limit).toList();
+        List<RestaurantEntity> restaurants = tierService.getRestaurantsByCuisinesAndSituationsAndLocationsWithPage(cuisines, situations, locations, null, true, page, limit).toList();
 
         if (restaurants.isEmpty()) {
             return new ResponseEntity<>(new ArrayList<>(), HttpStatus.OK);
@@ -144,7 +148,7 @@ public class TierApiController {
         // page 0부터 시작하게 수정
         page--;
         // DB 조회
-        List<RestaurantEntity> restaurants = restaurantApiService.getRestaurantsByCuisinesAndSituationsAndLocationsWithPage(cuisines, situations, locations, null, true, page, limit).toList();
+        List<RestaurantEntity> restaurants = tierService.getRestaurantsByCuisinesAndSituationsAndLocationsWithPage(cuisines, situations, locations, null, true, page, limit).toList();
 
         if (restaurants.isEmpty()) {
             return new ResponseEntity<>(new ArrayList<>(), HttpStatus.OK);
@@ -218,8 +222,8 @@ public class TierApiController {
         User user = userService.findUserById(userId);
 
         // 1. 음식 종류랑 위치로 식당 리스트 가져오기
-        List<RestaurantEntity> tieredRestaurants = restaurantApiService.getRestaurantsByCuisinesAndSituationsAndLocations(cuisines, situations, locations, 1, false);
-        List<RestaurantEntity> nonTieredRestaurants = restaurantApiService.getRestaurantsByCuisinesAndSituationsAndLocations(cuisines, situations, locations, -1, false);
+        List<RestaurantEntity> tieredRestaurants = tierService.getRestaurantsByCuisinesAndSituationsAndLocations(cuisines, situations, locations, 1, false);
+        List<RestaurantEntity> nonTieredRestaurants = tierService.getRestaurantsByCuisinesAndSituationsAndLocations(cuisines, situations, locations, -1, false);
 
         // 2. 상황으로 필터링하기
         List<RestaurantTierDTO> tieredRestaurantTierDTOs = tieredRestaurants.stream().map(restaurant ->
@@ -327,8 +331,8 @@ public class TierApiController {
         User user = userService.findUserById(userId);
 
         // 1. 음식 종류랑 위치로 식당 리스트 가져오기
-        List<RestaurantEntity> tieredRestaurants = restaurantApiService.getRestaurantsByCuisinesAndSituationsAndLocations(cuisines, situations, locations, 1, false);
-        List<RestaurantEntity> nonTieredRestaurants = restaurantApiService.getRestaurantsByCuisinesAndSituationsAndLocations(cuisines, situations, locations, -1, false);
+        List<RestaurantEntity> tieredRestaurants = tierService.getRestaurantsByCuisinesAndSituationsAndLocations(cuisines, situations, locations, 1, false);
+        List<RestaurantEntity> nonTieredRestaurants = tierService.getRestaurantsByCuisinesAndSituationsAndLocations(cuisines, situations, locations, -1, false);
 
         // 2. 상황으로 필터링하기
         List<RestaurantTierDTO> tieredRestaurantTierDTOs = tieredRestaurants.stream().map(restaurant ->
