+++ src/main/java/com/kustaurant/kustaurant/web/restaurant/TierWebController.java
@@ -6,12 +6,13 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.kustaurant.kustaurant.common.evaluation.service.port.EvaluationRepository;
 import com.kustaurant.kustaurant.common.restaurant.domain.dto.RestaurantTierDataClass;
 import com.kustaurant.kustaurant.common.restaurant.infrastructure.restaurant.RestaurantEntity;
+import com.kustaurant.kustaurant.common.restaurant.service.TierService;
 import com.kustaurant.kustaurant.global.webUser.CustomOAuth2UserService;
 import com.kustaurant.kustaurant.common.restaurant.argument_resolver.CuisineList;
 import com.kustaurant.kustaurant.common.restaurant.argument_resolver.LocationList;
 import com.kustaurant.kustaurant.common.restaurant.argument_resolver.SituationList;
 import com.kustaurant.kustaurant.common.evaluation.infrastructure.situation.SituationRepository;
-import com.kustaurant.kustaurant.api.restaurant.RestaurantApiService;
+import com.kustaurant.kustaurant.common.restaurant.service.RestaurantApiService;
 import com.kustaurant.kustaurant.common.user.infrastructure.User;
 import com.kustaurant.kustaurant.common.evaluation.service.EvaluationService;
 import jakarta.servlet.http.HttpServletRequest;
@@ -41,6 +42,7 @@ public class TierWebController {
     private final EvaluationService evaluationService;
     private final CustomOAuth2UserService customOAuth2UserService;
     //
+    private final TierService tierService;
     private final RestaurantApiService restaurantApiService;
 
     public static final Integer tierPageSize = 40;
@@ -94,7 +96,7 @@ public class TierWebController {
         // User 처리
         User user = customOAuth2UserService.getUserByPrincipal(principal);
         // restaurant list 처리
-        List<RestaurantEntity> tierRestaurants = restaurantApiService.getRestaurantsByCuisinesAndSituationsAndLocations(cuisines, situations, locations, null, true);
+        List<RestaurantEntity> tierRestaurants = tierService.getRestaurantsByCuisinesAndSituationsAndLocations(cuisines, situations, locations, null, true);
 
         List<RestaurantTierDataClass> restaurantsData = evaluationService.convertToTierDataClassList(tierRestaurants, user, true);
         Pageable pageable = PageRequest.of(page, tierPageSize);
